pyproject.toml
[tool.poetry]
name = "bina"
version = "0.1.0"
description = "Deterministic, rule-based static analysis tool for Python."
authors = ["Bina Team"]
readme = "README.md"
packages = [{include = "bina"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.1"
rich = "^13.0"
PyYAML = "^6.0"
requests = "^2.31"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
isort = "^5.0"
flake8 = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
bina = "bina.cli:main"
